+++ src/ATAP.Utilities.SystemParityMonitor.PowerShell/ATAP.Utilities.SystemParityMonitor.PowerShell.psm1
@@ -14,5 +14,6 @@ Export-ModuleMember -Function @(
   'Get-PeerPendingChanges',
   'Confirm-ParityChangeApplied',
   'Invoke-ParityAudit',
+  'Invoke-SharedDotNetToolConsumerProbe',
   'Compare-ParityAudits'
 )